+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add compatibility for Predis version 2.x
 - `LOAD ALL` option on Aggregate command (RediSearch `2.0.13`)
 - Profile command (RediSearch `2.2.0`) ([Issue#4])
+- Vector feature (RediSearch `2.4.0` / `2.4.2`)
 - (dev) Add `make clean` to remove all generated files
 - (dev) Add code coverage to `\MacFJA\RediSearch\Index`
 - (dev) Add integration test for document insertion